This was changed to a long

+++ src/utils/DateTime.java
@@ -163,7 +163,7 @@ public class DateTime {
    * @throws IllegalArgumentException if the interval was malformed.
    */
   public static final long parseDuration(final String duration) {
-    int interval;
+    long interval;
     int unit = 0;
     while (Character.isDigit(duration.charAt(unit))) {
       unit++;
